CRUD departments + little things fixed regarding adding in the database

+++ backend/src/main/java/org/example/licenta/services/UserService.java
@@ -6,6 +6,7 @@ import org.example.licenta.db.entities.enums.UserRoles;
 import org.example.licenta.db.repositories.AuthenticationRepository;
 import org.example.licenta.db.repositories.UserRepository;
 import org.example.licenta.dto.UserDto;
+import org.example.licenta.exceptions.UserAlreadyExistsException;
 import org.example.licenta.exceptions.UserNotFoundException;
 import org.example.licenta.mappers.AuthenticationMapper;
 import org.example.licenta.mappers.UserMapper;
@@ -34,56 +35,73 @@ public class UserService {
         this.authMapper = authMapper;
     }
 
-    public List<UserDto> getUsers() {
-        List<UserEntity> users = userRepository.findAll();
-        return users.stream().map(userMapper::toDto).collect(Collectors.toList());
+    public List<UserDto> getUsers() throws UserNotFoundException {
+        if (userRepository.findAll().isEmpty()) {
+            throw new UserNotFoundException("No users found");
+        }
+        else {
+            List<UserEntity> users = userRepository.findAll();
+            return users.stream().map(userMapper::toDto).collect(Collectors.toList());
+        }
     }
 
     public UserDto getUserById(String id) throws UserNotFoundException {
         Optional<UserEntity> user = userRepository.findById(id);
         if (user.isEmpty()) {
-                throw new UserNotFoundException("User not found");
+            throw new UserNotFoundException("User not found");
         }
         return userMapper.toDto(user.get());
     }
 
-    public void deleteUser(String id) {
-        userRepository.deleteById(id);
-        authRepository.deleteById(id);
+    public void deleteUser(String id) throws UserNotFoundException {
+        if (userRepository.findById(id).isEmpty()) {
+            throw new UserNotFoundException("User not found");
+        }
+        else {
+            userRepository.deleteById(id);
+            authRepository.deleteById(id);
+        }
     }
 
-    public void createUser(UserDto userDto) {
-        UserEntity userEntity = new UserEntity();
-        userEntity.setUserId(userDto.getUserId());
-        userEntity.setUserName(userDto.getUserName());
-        userEntity.setUserFirstName(userDto.getUserFirstName());
-        userEntity.setUserEmail(userDto.getUserEmail());
+    public void createUser(UserDto userDto) throws UserAlreadyExistsException {
+        if (userRepository.existsById(userDto.getUserId())) {
+            throw new UserAlreadyExistsException("User already exists");
+        }
+        else {
+            UserEntity userEntity = new UserEntity();
+            userEntity.setUserId(userDto.getUserId());
+            userEntity.setUserName(userDto.getUserName());
+            userEntity.setUserFirstName(userDto.getUserFirstName());
+            userEntity.setUserEmail(userDto.getUserEmail());
 //        TODO: Encode the password
-        userEntity.setUserPassword(userDto.getUserPassword());
-        userEntity.setUserRole(UserRoles.valueOf(userDto.getUserRole()));
+            userEntity.setUserPassword(userDto.getUserPassword());
+            userEntity.setUserRole(UserRoles.valueOf(userDto.getUserRole()));
 
 //        TODO: nu merge ca-i fk
 //        userEntity.setTeamEntity(userEntity.setTeamId("CAL"));
 //
-        userRepository.save(userEntity);
-        authRepository.save(authMapper.registered(userEntity));
+            userRepository.save(userEntity);
+            authRepository.save(authMapper.registered(userEntity));
+        }
     }
 
-    public UserDto updateUser(UserDto userDto, String id) {
+    public UserDto updateUser(UserDto userDto, String id) throws UserNotFoundException {
         Optional<UserEntity> user = userRepository.findById(id);
         if (user.isEmpty()) {
-            throw new RuntimeException("User not found");
+            throw new UserNotFoundException("User not found");
         }
-        user.get().setUserName(userDto.getUserName());
-        user.get().setUserFirstName(userDto.getUserFirstName());
-        user.get().setUserEmail(userDto.getUserEmail());
-        user.get().setUserPassword(userDto.getUserPassword());
-        user.get().setUserRole(UserRoles.valueOf(userDto.getUserRole()));
+        else {
+            user.get().setUserName(userDto.getUserName());
+            user.get().setUserFirstName(userDto.getUserFirstName());
+            user.get().setUserEmail(userDto.getUserEmail());
+            user.get().setUserPassword(userDto.getUserPassword());
+            user.get().setUserRole(UserRoles.valueOf(userDto.getUserRole()));
 //        TODO: Implement the team update
 //        user.get().setTeamEntity(userDto.getTeamEntity());
 //
-        userRepository.save(user.get());
-        authRepository.save(authMapper.registered(user.get()));
-        return userMapper.toDto(user.get());
+            userRepository.save(user.get());
+            authRepository.save(authMapper.registered(user.get()));
+            return userMapper.toDto(user.get());
+        }
     }
 }
